requirements.txt
torch
torchvision
Cython
matplotlib
tqdm
numpy
scipy
PyYAML
tensorboardX
tensorboard
scikit-learn
scikit-image
seaborn
pillow
opencv-contrib-python
einops
lpips
natsort
timm
openpyxl
kornia
xlwt==1.3.0
xlrd==1.2.0
pyiqa
rotary_embedding_torch
opencv-python>=4.8.0.76
opencv-python-headless>=4.5.5.64